requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
schedule==1.2.1
lxml==4.9.3 